pyproject.toml
[tool.poetry]
name = "delivery-service"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
